[dart2js] drop unnecessary modular data (2): non foreign nodes
The modular data saved references to every static invocation node, but only
needed this data for nodes representing one of the `JS*` foreign helpers.

This together with the previous change collectively reduce the total size of
the modular data by an average 8% on a small sample (which includes the entire
SDK).

Change-Id: I3080a1d7ba03643667cb50d7eadcc9fa0a065e6a
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/299921
Reviewed-by: Nate Biggs <natebiggs@google.com>