[dart2js] drop unnecessary data from modular data.
We seem to unnecessarily store all literals when all we need is to know
whether a literal of each kind is in use in the program.

I expect similar changes could help make the modular data more compact and
potentially make enough of a difference where a modular analysis may make
sense.

Change-Id: I045401d742b0467435da179af51c39c3ea08401f
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/299960
Reviewed-by: Nate Biggs <natebiggs@google.com>