[flow analysis] clean up public methods that call other public methods.
In a future CL, I plan to introduce some state tracking logic at the
top of each public method in `FlowAnalysisImpl` to verify that the
caller invokes flow analysis API methods in an order that respects
some important invariants. When I do that, it will be confusing if
there are some public methods in `FlowAnalysisImpl` that call other
public methods.

So, to avoid confusion in that future CL, this CL refactors the public
methods `lateInitializer_end` and `functionExpression_end` (which are
functionally equivalent) so that they call a common private method,
rather than one calling the other. Similarly, it refactors the public
methods `parenthesizedExpression` and `forwardExpression` (which are
functionally equivalent) so that they call a common private method,
rather than one calling the other.

Change-Id: I7b6d23f2db409c652575beb73bba2ea11b0fa90b
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/303200
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>