[infra] Fix utility used when bisecting CI failures
The bisection capability for failing builds on the CI is not much
used, and it is failing when updating the build information in Firestore.

A simple error in the script is causing the current error, so
this CL fixes a simple error in tools/bots/update_blamelists.dart.

Bug: https://github.com/dart-lang/sdk/issues/51978
Change-Id: Iba71e6115427145d20458d9e56f4cfa73fa232c9
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/302923
Auto-Submit: William Hesse <whesse@google.com>
Commit-Queue: Jonas Termansen <sortie@google.com>
Reviewed-by: Jonas Termansen <sortie@google.com>