Fix incorrect map indexing operation in flow analysis test.
Because of the permissive type of `Map.operator []`, I forgot to
update this test when I changed the key type of
`FlowModel.variableInfo` from `Variable?` to `int` (way back in
https://dart-review.googlesource.com/c/sdk/+/250120). The test passed
anyway because it was attempting to check that a certain key was not
present, so of course the test expectation was satisfied by passing in
a key that couldn't possibly exist in the map.

Fortunately the functionality in question had not regressed.

Change-Id: I9a8d6e75e4ca9419b7dae02d3ececa7b9a6ac0dd
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/305062
Commit-Queue: Konstantin Shcheglov <scheglov@google.com>
Auto-Submit: Paul Berry <paulberry@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>