feat: honour HttpOverrides in WebSocket.connect
The WebSocket implementation currently initializes the HttpClient it
uses in a static field. Consequently WebSocket.connect will not use any
custom HttpClient set using HttpOverrides.

This behaviour is changed to honour HttpOverrides by initialzing a new
client every time WebSocket.connect is called. The current behaviour for
setting the userAgent to be used for websocket requests is retained. If
a custom HttpClient is given in the connect call, the userAgent field is
not used. Furthermore the userAgent of the default client is only set if
the userAgent field of the WebSocket class was at least set once.