[analyzer] ContextRootImpl.isAnalyzed already have the resource, don't get it again
When analyzing rwf-materials (562 contexts) with a filled cache this
CL removes something like ~1.6 million calls to "isDirectorySync" which -
I think - reads from the filesystem.

On Linux I get this change (on Windows I'm guessing the change is bigger
as the filesystem is often slower):

Before the fix:
111662 ms
112053 ms
112176 ms

With this fix:
102829 ms
100934 ms
100991 ms

Difference at 95.0% confidence
        -10379 +/- 1780.48
        -9.26997% +/- 1.59023%
        (Student's t, pooled s = 785.532)

I.e. it makes it ~10 seconds or ~9% faster.

Change-Id: I42f54b2ff7510365c5a683c993b6b759b35bd152
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/305661
Commit-Queue: Jens Johansen <jensj@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>